code
This project help us in identification of the correct password. which is typed by the user .

It  includes  a microcontroller and 4X4 matrix keyboard to type the password .

This password will be compared by the password which is set through code , if it is correct then green LED will be glown ,else red LED will be  glown

By this ,we could able to know the status of access, whether the  access is denied or access is given.